+++ src/main/java/org/summerboot/jexpress/nio/grpc/GRPCClientConfig.java
@@ -28,6 +28,7 @@ import io.grpc.netty.shaded.io.netty.handler.ssl.SslContext;
 import io.grpc.netty.shaded.io.netty.handler.ssl.SslContextBuilder;
 import io.grpc.netty.shaded.io.netty.handler.ssl.SslProvider;
 import jakarta.annotation.Nullable;
+import org.summerboot.jexpress.boot.BootConstant;
 import org.summerboot.jexpress.boot.config.BootConfig;
 import org.summerboot.jexpress.boot.config.ConfigUtil;
 import org.summerboot.jexpress.boot.config.annotation.Config;
@@ -151,10 +152,10 @@ abstract public class GRPCClientConfig extends BootConfig {
     protected volatile KeyManagerFactory kmf;
 
     protected void generateTemplate_keystore(StringBuilder sb) {
-        sb.append(KEY_kmf_key + "=" + FILENAME_KEYSTORE + "\n");
-        sb.append(KEY_kmf_StorePwdKey + DEFAULT_DEC_VALUE);
-        sb.append(KEY_kmf_AliasKey + "=server3_4096.jexpress.org\n");
-        sb.append(KEY_kmf_AliasPwdKey + DEFAULT_DEC_VALUE);
+        sb.append(KEY_kmf_key + "=" + FILENAME_KEYSTORE + BootConstant.BR);
+        sb.append(KEY_kmf_StorePwdKey + DEFAULT_DEC_VALUE + BootConstant.BR);
+        sb.append(KEY_kmf_AliasKey + "=server3_4096.jexpress.org" + BootConstant.BR);
+        sb.append(KEY_kmf_AliasPwdKey + DEFAULT_DEC_VALUE + BootConstant.BR);
         generateTemplate = true;
     }
 
@@ -168,8 +169,8 @@ abstract public class GRPCClientConfig extends BootConfig {
     protected volatile TrustManagerFactory tmf;
 
     protected void generateTemplate_truststore(StringBuilder sb) {
-        sb.append(KEY_tmf_key + "=" + FILENAME_TRUSTSTORE_4CLIENT + "\n");
-        sb.append(KEY_tmf_StorePwdKey + DEFAULT_DEC_VALUE);
+        sb.append(KEY_tmf_key + "=" + FILENAME_TRUSTSTORE_4CLIENT + BootConstant.BR);
+        sb.append(KEY_tmf_StorePwdKey + DEFAULT_DEC_VALUE + BootConstant.BR);
         generateTemplate = true;
     }