+++ src/main/java/org/summerboot/jexpress/nio/server/NioConfig.java
@@ -118,10 +118,10 @@ public class NioConfig extends BootConfig {
     protected volatile KeyManagerFactory kmf = null;
 
     protected void generateTemplate_keystore(StringBuilder sb) {
-        sb.append(KEY_kmf_key + "=" + FILENAME_KEYSTORE + "\n");
-        sb.append(KEY_kmf_StorePwdKey + DEFAULT_DEC_VALUE);
-        sb.append(KEY_kmf_AliasKey + "=server1_2048.jexpress.org\n");
-        sb.append(KEY_kmf_AliasPwdKey + DEFAULT_DEC_VALUE);
+        sb.append(KEY_kmf_key + "=" + FILENAME_KEYSTORE + BootConstant.BR);
+        sb.append(KEY_kmf_StorePwdKey + DEFAULT_DEC_VALUE + BootConstant.BR);
+        sb.append(KEY_kmf_AliasKey + "=server1_2048.jexpress.org" + BootConstant.BR);
+        sb.append(KEY_kmf_AliasPwdKey + DEFAULT_DEC_VALUE + BootConstant.BR);
         generateTemplate = true;
     }
 
@@ -133,7 +133,7 @@ public class NioConfig extends BootConfig {
     protected volatile TrustManagerFactory tmf = null;
 
     //    protected void generateTemplate_truststore(StringBuilder sb) {
-//        sb.append(KEY_tmf_key + "="+FILENAME_TRUSTSTORE_4SERVER+"\n");
+//        sb.append(KEY_tmf_key + "="+FILENAME_TRUSTSTORE_4SERVER + BootConstant.BR);
 //        sb.append(KEY_tmf_StorePwdKey + DEFAULT_DEC_VALUE);
 //        generateTemplate = true;
 //    }
@@ -412,7 +412,7 @@ public class NioConfig extends BootConfig {
     protected HttpHeaders serverDefaultResponseHeaders;
 
     protected void generateTemplate_ResponseHeaders(StringBuilder sb) {
-        sb.append("#").append(HEADER_SERVER_RESPONSE).append("response_header_name=response_header_value\n");
+        sb.append("#").append(HEADER_SERVER_RESPONSE).append("response_header_name=response_header_value" + BootConstant.BR);
     }
 
     public HttpHeaders getServerDefaultResponseHeaders() {